+++ apps/wpcom-smart-dictation/wpcom-smart-dictation.jsx
@@ -1,15 +1,14 @@
 /* global __i18n_text_domain__ */
 import './config';
-import { LiveAIAssistant, MicIcon } from '@automattic/help-center';
+import { LiveAIAssistant, MicIcon } from '@automattic/wpcom-smart-dictation';
 import { PluginSidebar, PluginSidebarMoreMenuItem } from '@wordpress/editor';
 import { __ } from '@wordpress/i18n';
 import { registerPlugin } from '@wordpress/plugins';
-import './help-center.scss';
 
 const SMART_DICTATION_SIDEBAR_NAME = 'wpcom-smart-dictation';
 const MicrophoneIcon = () => <MicIcon animated size={ 20 } />;
 
-function JetpackSmartDictationPlugin() {
+function WpcomSmartDictationPlugin() {
 	return (
 		<>
 			<PluginSidebarMoreMenuItem target={ SMART_DICTATION_SIDEBAR_NAME } icon={ MicrophoneIcon }>
@@ -28,7 +27,7 @@ function JetpackSmartDictationPlugin() {
 	);
 }
 
-registerPlugin( 'jetpack-live-ai-assistant', {
+registerPlugin( 'wpcom-smart-dictation', {
 	icon: MicrophoneIcon,
-	render: () => <JetpackSmartDictationPlugin />,
+	render: () => <WpcomSmartDictationPlugin />,
 } );
