+++ packages/help-center/src/index.ts
@@ -8,5 +8,3 @@ export { default as HelpCenterInlineButton } from './components/help-center-inli
 export { default as Mail } from './icons/mail';
 export { useHelpSearchQuery } from './hooks/use-help-search-query';
 export { HelpCenterSupportGuides } from './components/help-center-support-guides';
-export { LiveAIAssistant } from './live-ai-assistant';
-export { MicIcon } from './live-ai-assistant/components/mic-icon';
