+++ packages/wpcom-smart-dictation/src/index.ts
@@ -1,0 +1,2 @@
+export { default, LiveAIAssistant } from './live-ai-assistant';
+export { MicIcon } from './live-ai-assistant/components/mic-icon';
