+++ packages/wpcom-smart-dictation/src/live-ai-assistant/components/mic-icon.tsx
