+++ packages/wpcom-smart-dictation/src/live-ai-assistant/image-picker-modal.tsx
