+++ packages/wpcom-smart-dictation/src/live-ai-assistant/realtime-session/constants.ts
@@ -1,0 +1,8 @@
+export const DEFAULT_MODEL = 'gpt-realtime';
+/** Proxied by WP.com REST to OpenAI `/v1/realtime/client_secrets`. */
+export const REALTIME_CLIENT_SECRETS_PATH = '/ai-api-proxy/v1/realtime/client_secrets';
+export const OPENAI_REALTIME_URL = 'https://api.openai.com/v1/realtime/calls';
+export const MAX_TOOL_EVENTS = 40;
+export const PROGRAMMATIC_SEND_TIMEOUT_MS = 15000;
+export const TOOL_ERROR_MESSAGE_MAX_LENGTH = 500;
+export const GET_USER_MEDIA_TIMEOUT_MS = 30_000;
