+++ packages/wpcom-smart-dictation/src/live-ai-assistant/realtime-session/index.ts
@@ -1,0 +1,8 @@
+export { useRealtimeSession } from './use-realtime-session';
+export type {
+	RealtimeStatus,
+	RealtimeToolEvent,
+	RealtimeTranscriptEntry,
+	UseRealtimeSessionOptions,
+	UseRealtimeSessionResult,
+} from './types';
