+++ packages/wpcom-smart-dictation/src/live-ai-assistant/realtime-session/use-realtime-session.ts
@@ -1,0 +1,579 @@
+import { recordTracksEvent } from '@automattic/calypso-analytics';
+import { useCallback, useEffect, useRef, useState } from 'react';
+import { fetchEphemeralKey } from './client-secret';
+import {
+	DEFAULT_MODEL,
+	GET_USER_MEDIA_TIMEOUT_MS,
+	MAX_TOOL_EVENTS,
+	OPENAI_REALTIME_URL,
+	PROGRAMMATIC_SEND_TIMEOUT_MS,
+} from './constants';
+import { getErrorMessage, getErrorName } from './errors';
+import { realtimeToolDefinitions } from './tool-definitions';
+import { executeRealtimeToolCalls } from './tool-runner';
+import { assistantTurnEntryId, upsertEntry } from './transcript';
+import type {
+	RealtimeStatus,
+	RealtimeToolEvent,
+	RealtimeTranscriptEntry,
+	UseRealtimeSessionOptions,
+	UseRealtimeSessionResult,
+} from './types';
+
+declare global {
+	interface Window {
+		/**
+		 * TEMPORARY: inject user text into the active Smart Dictation realtime
+		 * session (e.g. from the devtools console). Remove when no longer needed.
+		 */
+		sendToDictation?: ( text: string ) => Promise< void >;
+	}
+}
+
+/**
+ * Manages an OpenAI Realtime API session over WebRTC. Microphone captures the
+ * only user modality (speech); assistant turns stream as text (no synthesized
+ * speech). Data-channel events carry transcripts and tool callbacks.
+ */
+export function useRealtimeSession( options: UseRealtimeSessionOptions ): UseRealtimeSessionResult {
+	const { model = DEFAULT_MODEL, instructions } = options;
+
+	const [ status, setStatus ] = useState< RealtimeStatus >( 'idle' );
+	const [ error, setError ] = useState< string | null >( null );
+	const [ isMuted, setIsMuted ] = useState( false );
+	const [ transcript, setTranscript ] = useState< RealtimeTranscriptEntry[] >( [] );
+	const [ toolEvents, setToolEvents ] = useState< RealtimeToolEvent[] >( [] );
+	const [ imagePickerState, setImagePickerState ] = useState<
+		import('../image-picker-modal').ImagePickerState
+	>( {
+		isOpen: false,
+		images: [],
+		selectedNumber: null,
+		purpose: 'block',
+	} );
+
+	useEffect( () => {
+		const onUpdate = () => {
+			const s = window.__dictationImagePicker;
+			if ( s ) {
+				setImagePickerState( { ...s } );
+			}
+		};
+		window.addEventListener( 'dictation-image-picker-update', onUpdate );
+		return () => window.removeEventListener( 'dictation-image-picker-update', onUpdate );
+	}, [] );
+
+	const peerConnectionRef = useRef< RTCPeerConnection | null >( null );
+	const dataChannelRef = useRef< RTCDataChannel | null >( null );
+	const localStreamRef = useRef< MediaStream | null >( null );
+	const transcriptCountRef = useRef( 0 );
+	const toolEventsCountRef = useRef( 0 );
+	const sessionStartedAtRef = useRef< number | null >( null );
+	const hasTrackedSessionStartRef = useRef( false );
+	const programmaticSendQueueRef = useRef< Promise< void > >( Promise.resolve() );
+	// Tracks whether a response is currently being generated by the server.
+	const activeResponseRef = useRef( false );
+	// Set when safeCreateResponse() is called while activeResponseRef is true.
+	const pendingResponseCreateRef = useRef( false );
+
+	const recordSessionEnded = useCallback( ( reason: string ) => {
+		if ( ! hasTrackedSessionStartRef.current || ! sessionStartedAtRef.current ) {
+			return;
+		}
+		const durationMs = Date.now() - sessionStartedAtRef.current;
+		recordTracksEvent( 'calypso_smart_dictation_ended', {
+			reason,
+			duration_ms: durationMs,
+			duration_seconds: Math.round( durationMs / 1000 ),
+		} );
+		hasTrackedSessionStartRef.current = false;
+		sessionStartedAtRef.current = null;
+	}, [] );
+
+	const cleanup = useCallback( () => {
+		activeResponseRef.current = false;
+		pendingResponseCreateRef.current = false;
+		programmaticSendQueueRef.current = Promise.resolve();
+
+		try {
+			dataChannelRef.current?.close();
+		} catch {}
+		dataChannelRef.current = null;
+
+		try {
+			peerConnectionRef.current?.getSenders().forEach( ( sender ) => {
+				try {
+					sender.track?.stop();
+				} catch {}
+			} );
+			peerConnectionRef.current?.close();
+		} catch {}
+		peerConnectionRef.current = null;
+
+		localStreamRef.current?.getTracks().forEach( ( track ) => track.stop() );
+		localStreamRef.current = null;
+	}, [] );
+
+	useEffect( () => {
+		return () => cleanup();
+	}, [ cleanup ] );
+
+	useEffect( () => {
+		transcriptCountRef.current = transcript.length;
+	}, [ transcript ] );
+
+	useEffect( () => {
+		toolEventsCountRef.current = toolEvents.length;
+	}, [ toolEvents ] );
+
+	/**
+	 * Sends a response.create event, but only if no response is currently in
+	 * flight. If one is, queues a single deferred send via
+	 * pendingResponseCreateRef which gets drained when the active response ends.
+	 */
+	const safeCreateResponse = useCallback( () => {
+		const dc = dataChannelRef.current;
+		if ( ! dc || dc.readyState !== 'open' ) {
+			return;
+		}
+		if ( activeResponseRef.current ) {
+			pendingResponseCreateRef.current = true;
+			return;
+		}
+		activeResponseRef.current = true;
+		pendingResponseCreateRef.current = false;
+		dc.send(
+			JSON.stringify( {
+				type: 'response.create',
+				response: {
+					output_modalities: [ 'text' ],
+				},
+			} )
+		);
+	}, [] );
+
+	const flushPendingResponseCreate = useCallback( () => {
+		if ( ! pendingResponseCreateRef.current ) {
+			return;
+		}
+		pendingResponseCreateRef.current = false;
+		safeCreateResponse();
+	}, [ safeCreateResponse ] );
+
+	const handleToolCalls = useCallback(
+		async ( event: { response?: { output?: unknown[] } } ) => {
+			const dc = dataChannelRef.current;
+			if ( ! dc || dc.readyState !== 'open' ) {
+				return;
+			}
+
+			const didExecuteTools = await executeRealtimeToolCalls( {
+				event,
+				onToolEvent: ( toolEvent ) => {
+					setToolEvents( ( prev ) => {
+						const next = prev.concat( toolEvent );
+						return next.length > MAX_TOOL_EVENTS
+							? next.slice( next.length - MAX_TOOL_EVENTS )
+							: next;
+					} );
+				},
+				sendFunctionCallOutput: ( callId, result ) => {
+					dc.send(
+						JSON.stringify( {
+							type: 'conversation.item.create',
+							item: {
+								type: 'function_call_output',
+								call_id: callId,
+								output: JSON.stringify( result ),
+							},
+						} )
+					);
+				},
+			} );
+
+			if ( didExecuteTools ) {
+				safeCreateResponse();
+			}
+		},
+		[ safeCreateResponse ]
+	);
+
+	const handleServerEvent = useCallback(
+		( event: unknown ) => {
+			if ( ! event || typeof event !== 'object' ) {
+				return;
+			}
+			const evt = event as { type?: string; [ key: string ]: unknown };
+
+			switch ( evt.type ) {
+				case 'conversation.item.input_audio_transcription.delta':
+				case 'conversation.item.input_audio_transcription.completed': {
+					const itemId = ( evt.item_id as string ) || 'user-latest';
+					const delta = ( evt.delta as string ) || ( evt.transcript as string ) || '';
+					const isFinal = evt.type.endsWith( 'completed' );
+					setTranscript( ( prev ) => upsertEntry( prev, itemId, 'user', delta, isFinal ) );
+					break;
+				}
+				case 'response.text.delta':
+				case 'response.output_text.delta':
+				case 'response.audio_transcript.delta':
+				case 'response.output_audio_transcript.delta':
+				case 'response.text.done':
+				case 'response.output_text.done':
+				case 'response.audio_transcript.done':
+				case 'response.output_audio_transcript.done': {
+					const itemId = assistantTurnEntryId( evt );
+					const isDone =
+						evt.type === 'response.text.done' ||
+						evt.type === 'response.output_text.done' ||
+						evt.type === 'response.audio_transcript.done' ||
+						evt.type === 'response.output_audio_transcript.done';
+					let delta = '';
+					if ( isDone ) {
+						if ( typeof evt.text === 'string' ) {
+							delta = evt.text;
+						} else if ( typeof evt.transcript === 'string' ) {
+							delta = evt.transcript;
+						}
+					} else if ( typeof evt.delta === 'string' ) {
+						delta = evt.delta;
+					}
+					setTranscript( ( prev ) => upsertEntry( prev, itemId, 'assistant', delta, isDone ) );
+					break;
+				}
+				case 'response.created': {
+					activeResponseRef.current = true;
+					break;
+				}
+				case 'response.done':
+				case 'response.cancelled':
+				case 'response.failed': {
+					activeResponseRef.current = false;
+					if ( evt.type === 'response.done' ) {
+						void handleToolCalls( evt as { response?: { output?: unknown[] } } );
+					}
+					flushPendingResponseCreate();
+					break;
+				}
+				case 'error': {
+					const errorObj = evt.error as { code?: string; message?: string } | undefined;
+					const message = errorObj?.message || 'Realtime session error';
+					const isActiveResponseConflict =
+						errorObj?.code === 'conversation_already_has_active_response' ||
+						/already\s+has\s+an?\s+active\s+response/i.test( message );
+					if ( isActiveResponseConflict ) {
+						activeResponseRef.current = true;
+						pendingResponseCreateRef.current = true;
+						break;
+					}
+					setError( message );
+					setStatus( 'error' );
+					recordSessionEnded( 'server_error' );
+					break;
+				}
+				default:
+					break;
+			}
+		},
+		[ flushPendingResponseCreate, handleToolCalls, recordSessionEnded ]
+	);
+
+	const start = useCallback( async () => {
+		if (
+			status === 'active' ||
+			status === 'connecting' ||
+			status === 'requesting-token' ||
+			status === 'requesting-mic'
+		) {
+			return;
+		}
+
+		setError( null );
+		setTranscript( [] );
+		setToolEvents( [] );
+		sessionStartedAtRef.current = Date.now();
+		hasTrackedSessionStartRef.current = true;
+		recordTracksEvent( 'calypso_smart_dictation_started' );
+
+		let hasRequestedMicrophone = false;
+
+		try {
+			await assertMicrophonePermission();
+
+			setStatus( 'requesting-token' );
+			const ephemeralKey = await fetchEphemeralKey( {
+				model,
+				instructions,
+			} );
+
+			setStatus( 'requesting-mic' );
+			recordTracksEvent( 'calypso_smart_dictation_microphone_permission_requested' );
+			hasRequestedMicrophone = true;
+			const localStream = await getUserMediaWithTimeout();
+			recordTracksEvent( 'calypso_smart_dictation_microphone_permission_granted' );
+			localStreamRef.current = localStream;
+
+			setStatus( 'connecting' );
+			const pc = new RTCPeerConnection();
+			peerConnectionRef.current = pc;
+
+			localStream.getTracks().forEach( ( track ) => pc.addTrack( track, localStream ) );
+
+			const dataChannel = pc.createDataChannel( 'oai-events' );
+			dataChannelRef.current = dataChannel;
+
+			dataChannel.addEventListener( 'open', () => {
+				setStatus( 'active' );
+				dataChannel.send(
+					JSON.stringify( {
+						type: 'session.update',
+						session: {
+							type: 'realtime',
+							output_modalities: [ 'text' ],
+							instructions,
+							tools: realtimeToolDefinitions,
+							tool_choice: 'auto',
+							audio: {
+								input: {
+									transcription: { model: 'whisper-1' },
+									turn_detection: { type: 'server_vad' },
+								},
+							},
+						},
+					} )
+				);
+			} );
+
+			dataChannel.addEventListener( 'message', ( event ) => {
+				try {
+					handleServerEvent( JSON.parse( event.data ) );
+				} catch {
+					// Non-JSON events are ignored.
+				}
+			} );
+
+			dataChannel.addEventListener( 'close', () => {
+				recordSessionEnded( 'data_channel_closed' );
+				setStatus( ( prev ) => ( prev === 'error' ? prev : 'idle' ) );
+			} );
+
+			pc.addEventListener( 'connectionstatechange', () => {
+				if (
+					pc.connectionState === 'failed' ||
+					pc.connectionState === 'disconnected' ||
+					pc.connectionState === 'closed'
+				) {
+					recordSessionEnded( pc.connectionState );
+					setStatus( ( prev ) => ( prev === 'error' ? prev : 'idle' ) );
+				}
+			} );
+
+			const offer = await pc.createOffer();
+			await pc.setLocalDescription( offer );
+
+			const sdpResponse = await fetch(
+				`${ OPENAI_REALTIME_URL }?model=${ encodeURIComponent( model ) }`,
+				{
+					method: 'POST',
+					body: offer.sdp,
+					headers: {
+						Authorization: `Bearer ${ ephemeralKey }`,
+						'Content-Type': 'application/sdp',
+					},
+				}
+			);
+
+			if ( ! sdpResponse.ok ) {
+				const errText = await sdpResponse.text().catch( () => '' );
+				let detail = errText;
+				try {
+					detail = JSON.parse( errText )?.error?.message ?? errText;
+				} catch {}
+				throw new Error(
+					`Realtime SDP exchange failed (${ sdpResponse.status })${ detail ? ': ' + detail : '' }`
+				);
+			}
+
+			const answerSdp = await sdpResponse.text();
+			await pc.setRemoteDescription( { type: 'answer', sdp: answerSdp } );
+		} catch ( err ) {
+			const message = err instanceof Error ? err.message : 'Unknown error starting session';
+			if ( hasRequestedMicrophone ) {
+				recordTracksEvent( 'calypso_smart_dictation_microphone_permission_failed', {
+					error_name: getErrorName( err ),
+					error_message: getErrorMessage( err ),
+				} );
+			}
+			setError( message );
+			setStatus( 'error' );
+			recordSessionEnded( 'error' );
+			cleanup();
+		}
+	}, [ cleanup, handleServerEvent, instructions, model, recordSessionEnded, status ] );
+
+	const stop = useCallback( () => {
+		recordSessionEnded( 'user_stop' );
+		setStatus( 'ending' );
+		cleanup();
+		setStatus( 'idle' );
+	}, [ cleanup, recordSessionEnded ] );
+
+	const toggleMute = useCallback( () => {
+		const stream = localStreamRef.current;
+		if ( ! stream ) {
+			return;
+		}
+		const nextMuted = ! isMuted;
+		stream.getAudioTracks().forEach( ( track ) => {
+			track.enabled = ! nextMuted;
+		} );
+		setIsMuted( nextMuted );
+	}, [ isMuted ] );
+
+	const sendText = useCallback(
+		( text: string ): Promise< void > => {
+			const trimmed = typeof text === 'string' ? text.trim() : '';
+			if ( ! trimmed.length ) {
+				return Promise.resolve();
+			}
+
+			const run = async () => {
+				const dc = dataChannelRef.current;
+				if ( ! dc || dc.readyState !== 'open' ) {
+					// eslint-disable-next-line no-console -- TEMPORARY dev hook
+					console.warn(
+						'[sendToDictation] No open realtime data channel (start dictation first).'
+					);
+					return;
+				}
+
+				const startTranscriptCount = transcriptCountRef.current;
+				const startToolEventsCount = toolEventsCountRef.current;
+
+				const itemId = `programmatic-user-${ Date.now() }`;
+				setTranscript( ( prev ) => upsertEntry( prev, itemId, 'user', trimmed, true ) );
+
+				dc.send(
+					JSON.stringify( {
+						type: 'conversation.item.create',
+						item: {
+							type: 'message',
+							role: 'user',
+							content: [ { type: 'input_text', text: trimmed } ],
+						},
+					} )
+				);
+				safeCreateResponse();
+
+				await new Promise< void >( ( resolve ) => {
+					const startedAt = Date.now();
+					const poll = () => {
+						const timedOut = Date.now() - startedAt >= PROGRAMMATIC_SEND_TIMEOUT_MS;
+						const hasProgress =
+							transcriptCountRef.current > startTranscriptCount ||
+							toolEventsCountRef.current > startToolEventsCount;
+						const isSettled = ! activeResponseRef.current && ! pendingResponseCreateRef.current;
+						if ( timedOut || ( hasProgress && isSettled ) ) {
+							resolve();
+							return;
+						}
+						window.setTimeout( poll, 100 );
+					};
+					poll();
+				} );
+			};
+
+			const queued = programmaticSendQueueRef.current.then( run );
+			programmaticSendQueueRef.current = queued.catch( () => undefined );
+			return queued;
+		},
+		[ safeCreateResponse ]
+	);
+
+	const sendEvent = useCallback( ( eventName: string, details?: string ) => {
+		void eventName;
+		void details;
+	}, [] );
+
+	useEffect( () => {
+		const fn: ( text: string ) => Promise< void > = ( text ) => {
+			return sendText( text );
+		};
+		window.sendToDictation = fn;
+		return () => {
+			if ( window.sendToDictation === fn ) {
+				delete window.sendToDictation;
+			}
+		};
+	}, [ sendText ] );
+
+	return {
+		status,
+		error,
+		isMuted,
+		transcript,
+		toolEvents,
+		imagePickerState,
+		start,
+		stop,
+		toggleMute,
+		sendText,
+		sendEvent,
+	};
+}
+
+async function assertMicrophonePermission(): Promise< void > {
+	if ( ! navigator.permissions ) {
+		recordTracksEvent( 'calypso_smart_dictation_microphone_permission_checked', {
+			supported: false,
+			state: 'unsupported',
+		} );
+		return;
+	}
+
+	let permissionState: PermissionState;
+	try {
+		const permissionStatus = await navigator.permissions.query( {
+			name: 'microphone' as PermissionName,
+		} );
+		permissionState = permissionStatus.state;
+		recordTracksEvent( 'calypso_smart_dictation_microphone_permission_checked', {
+			supported: true,
+			state: permissionState,
+		} );
+	} catch ( permErr ) {
+		recordTracksEvent( 'calypso_smart_dictation_microphone_permission_check_failed', {
+			error_name: getErrorName( permErr ),
+			error_message: getErrorMessage( permErr ),
+		} );
+		throw new Error(
+			'Microphone access is blocked. Please allow microphone access in your browser settings and try again.'
+		);
+	}
+
+	if ( permissionState === 'denied' ) {
+		recordTracksEvent( 'calypso_smart_dictation_microphone_permission_blocked', {
+			source: 'permissions_api',
+		} );
+		throw new Error(
+			'Microphone access is blocked. Please allow microphone access in your browser settings and try again.'
+		);
+	}
+}
+
+function getUserMediaWithTimeout(): Promise< MediaStream > {
+	return Promise.race( [
+		navigator.mediaDevices.getUserMedia( { audio: true } ),
+		new Promise< never >( ( _, reject ) =>
+			setTimeout(
+				() =>
+					reject(
+						new Error(
+							'Microphone request timed out. Check for a browser permission prompt near the address bar.'
+						)
+					),
+				GET_USER_MEDIA_TIMEOUT_MS
+			)
+		),
+	] );
+}
