+++ packages/wpcom-smart-dictation/src/live-ai-assistant/tools/image-picker-tool.ts
