+++ packages/wpcom-smart-dictation/src/live-ai-assistant/use-realtime-session.ts
@@ -1,0 +1,8 @@
+export { useRealtimeSession } from './realtime-session';
+export type {
+	RealtimeStatus,
+	RealtimeToolEvent,
+	RealtimeTranscriptEntry,
+	UseRealtimeSessionOptions,
+	UseRealtimeSessionResult,
+} from './realtime-session';
