+++ packages/wpcom-smart-dictation/src/types.d.ts
@@ -1,0 +1,2 @@
+declare module '*.mp4';
+declare module '*.scss';
