+++ packages/wpcom-smart-dictation/tsconfig.json
@@ -1,0 +1,13 @@
+{
+	"extends": "@automattic/calypso-typescript-config/ts-package.json",
+	"compilerOptions": {
+		"allowJs": true,
+		"checkJs": false,
+		"declarationDir": "dist/types",
+		"outDir": "dist/esm",
+		"rootDir": "src"
+	},
+	"include": [ "src", "types" ],
+	"exclude": [ "**/test/*" ],
+	"references": [ { "path": "../calypso-analytics" }, { "path": "../i18n-utils" } ]
+}
